requirements.txt
asgiref==3.5.0
Django==4.0.3
django-debug-toolbar==3.2.4
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
Pillow==9.1.0
python-decouple==3.6
sqlparse==0.4.2
